+++ netbox_contract/models.py
@@ -9,6 +9,8 @@ from django.utils.translation import gettext_lazy as _
 from netbox.models import NetBoxModel
 from netbox.models.features import ContactsMixin
 from utilities.choices import ChoiceSet
+from dcim.choices import DeviceStatusChoices, SiteStatusChoices
+from virtualization.choices import VirtualMachineStatusChoices
 
 
 class StatusChoices(ChoiceSet):
@@ -86,13 +88,9 @@ class AccountingDimension(NetBoxModel):
 
     def get_status_color(self):
         return AccountingDimensionStatusChoices.colors.get(self.status)
-    
+
     class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=['name', 'value'], name='unique_accounting_dimension'
-            )
-        ]
+        constraints = [models.UniqueConstraint(fields=['name', 'value'], name='unique_accounting_dimension')]
         ordering = ('name', 'value')
         verbose_name = _('accounting dimension')
         verbose_name_plural = _('accounting dimensions')
@@ -117,9 +115,7 @@ class ServiceProvider(ContactsMixin, NetBoxModel):
 
 
 class ContractAssignment(NetBoxModel):
-    content_type = models.ForeignKey(
-        to=ContentType, on_delete=models.CASCADE, verbose_name=_('content type')
-    )
+    content_type = models.ForeignKey(to=ContentType, on_delete=models.CASCADE, verbose_name=_('content type'))
     object_id = models.PositiveBigIntegerField(verbose_name=_('object ID'))
     content_object = GenericForeignKey(ct_field='content_type', fk_field='object_id')
     contract = models.ForeignKey(
@@ -141,6 +137,18 @@ class ContractAssignment(NetBoxModel):
     def get_absolute_url(self):
         return reverse('plugins:netbox_contract:contractassignment', args=[self.pk])
 
+    def get_contract__status_color(self):
+        return StatusChoices.colors.get(self.contract.status)
+
+    def get_content_object__status_color(self):
+        STATUS_MAPPING = {
+            'virtualmachine': VirtualMachineStatusChoices.colors,
+            'device': DeviceStatusChoices.colors,
+            'site': SiteStatusChoices.colors,
+        }
+        status_colors = STATUS_MAPPING.get(self.content_type.model, StatusChoices.colors)
+        return status_colors.get(self.content_object.status)
+
 
 class Contract(NetBoxModel):
     name = models.CharField(max_length=100, verbose_name=_('name'))
@@ -157,12 +165,9 @@ class Contract(NetBoxModel):
     external_partie_object = GenericForeignKey(
         ct_field='external_partie_object_type', fk_field='external_partie_object_id'
     )
-    external_reference = models.CharField(
-        max_length=100, blank=True, null=True, verbose_name=_('external reference')
-    )
-    internal_partie = models.CharField(
-        max_length=50, choices=InternalEntityChoices, verbose_name=_('internal partie')
-    )
+    external_partie_object.editable = True
+    external_reference = models.CharField(max_length=100, blank=True, null=True, verbose_name=_('external reference'))
+    internal_partie = models.CharField(max_length=50, choices=InternalEntityChoices, verbose_name=_('internal partie'))
     tenant = models.ForeignKey(
         to='tenancy.Tenant',
         on_delete=models.PROTECT,
@@ -220,9 +225,7 @@ class Contract(NetBoxModel):
         null=True,
         help_text=_('Use either this field of the yearly recuring cost field'),
     )
-    nrc = models.DecimalField(
-        verbose_name=_('none recuring cost'), default=0, max_digits=10, decimal_places=2
-    )
+    nrc = models.DecimalField(verbose_name=_('none recuring cost'), default=0, max_digits=10, decimal_places=2)
     invoice_frequency = models.IntegerField(
         help_text=_('The frequency of invoices in month'),
         default=1,
@@ -248,12 +251,11 @@ class Contract(NetBoxModel):
 
     def get_status_color(self):
         return StatusChoices.colors.get(self.status)
+
     class Meta:
         ordering = ('name',)
         indexes = [
-            models.Index(
-                fields=['external_partie_object_type', 'external_partie_object_id']
-            ),
+            models.Index(fields=['external_partie_object_type', 'external_partie_object_id']),
         ]
         verbose_name = _('contract')
         verbose_name_plural = _('contracts')
@@ -276,12 +278,8 @@ class Invoice(NetBoxModel):
         help_text=_('Wether this invoice is a template or not'),
     )
     date = models.DateField(blank=True, null=True, verbose_name=_('date'))
-    contracts = models.ManyToManyField(
-        Contract, related_name='invoices', blank=True, verbose_name=_('contracts')
-    )
-    period_start = models.DateField(
-        blank=True, null=True, verbose_name=_('period start')
-    )
+    contracts = models.ManyToManyField(Contract, related_name='invoices', blank=True, verbose_name=_('contracts'))
+    period_start = models.DateField(blank=True, null=True, verbose_name=_('period start'))
     period_end = models.DateField(blank=True, null=True, verbose_name=_('period end'))
     currency = models.CharField(
         max_length=3,
@@ -289,9 +287,7 @@ class Invoice(NetBoxModel):
         default=CURRENCY_DEFAULT,
         verbose_name=_('currency'),
     )
-    amount = models.DecimalField(
-        max_digits=10, decimal_places=2, verbose_name=_('amount')
-    )
+    amount = models.DecimalField(max_digits=10, decimal_places=2, verbose_name=_('amount'))
     documents = models.URLField(
         blank=True,
         verbose_name=_('documents'),
@@ -331,9 +327,7 @@ class InvoiceLine(NetBoxModel):
         default=CURRENCY_DEFAULT,
         verbose_name=_('currency'),
     )
-    amount = models.DecimalField(
-        max_digits=10, decimal_places=2, verbose_name=_('amount')
-    )
+    amount = models.DecimalField(max_digits=10, decimal_places=2, verbose_name=_('amount'))
     accounting_dimensions = models.ManyToManyField(
         AccountingDimension, blank=True, verbose_name=_('accounting dimensions')
     )
@@ -355,14 +349,8 @@ class InvoiceLine(NetBoxModel):
         is_new = not bool(self.pk)
         if is_new:
             if amount > (invoice.amount - invoice.total_invoicelines_amount):
-                raise ValidationError(
-                    'Sum of invoice line amount greater than invoice amount'
-                )
+                raise ValidationError('Sum of invoice line amount greater than invoice amount')
         else:
             previous_amount = self.__class__.objects.get(pk=self.pk).amount
-            if amount > (
-                invoice.amount - invoice.total_invoicelines_amount + previous_amount
-            ):
-                raise ValidationError(
-                    'Sum of invoice line amount greater than invoice amount'
-                )
+            if amount > (invoice.amount - invoice.total_invoicelines_amount + previous_amount):
+                raise ValidationError('Sum of invoice line amount greater than invoice amount')