+++ src/netbox_contract/constants.py
@@ -1,6 +1,0 @@
-from django.db.models import Q
-
-SERVICE_PROVIDER_MODELS = Q(
-    Q(app_label='circuits', model__in=('provider',))
-    | Q(app_label='netbox_contract', model__in=('serviceprovider',))
-)